Explicitly use find_package in Config mode.
I'm not completely sure why this is necessary as the documentation I can find
suggests that the default is for cmake to first look for packages using
FindTinyXML.cmake and then automatically fallback to "Config mode".

When building on Windows, in both CI and development generating fails with
```
-- Could NOT find TinyXML (missing:  TinyXML_LIBRARY TinyXML_INCLUDE_DIR)
```
unless a second find_package directive with the `CONFIG` option is set.