Try to find Config package in TinyXML Module file.
Seeing Dirk bring in this line from the CMake docs out of context

> If no module is found and the MODULE option is not given the
> command proceeds to Config mode.

Made me realize that "If no module is found" might mean "If no module
file is found" rather than "If no package is found with the module
files." which was my assumption from reading the docs previously.

Using tinyxml_vendor declaratively is nice for packages that rely on
it but it means that the entire burden of coordinating lies within the
package. I think it'd be easier if packages tried to resolve TinyXML
and only pulled in tinyxml_vendor as a last resort. It would simplify
tinyxml_vendor and cut down on special cases. But if this attempt
is successful, it might be fine.